Parametrize authz test, move some logic to layman.common.rest
