--- src/layman/authz/util.py
@@ -1,22 +1,0 @@
-from layman import settings
-
-
-def setup_patch_access_rights(request_form, kwargs):
-    for type in ['read', 'write']:
-        if request_form.get('access_rights.' + type):
-            kwargs['access_rights'] = kwargs.get('access_rights', dict())
-            access_rights = list({x.strip() for x in request_form['access_rights.' + type].split(',')})
-            kwargs['access_rights'][type] = access_rights
-
-
-def setup_post_access_rights(request_form, kwargs, actor_name):
-    kwargs['access_rights'] = dict()
-    for type in ['read', 'write']:
-        if not request_form.get('access_rights.' + type):
-            if actor_name:
-                access_rights = [actor_name]
-            else:
-                access_rights = [settings.RIGHTS_EVERYONE_ROLE]
-        else:
-            access_rights = list({x.strip() for x in request_form['access_rights.' + type].split(',')})
-        kwargs['access_rights'][type] = access_rights
