+++ src/layman/map/rest_maps.py
@@ -4,6 +4,7 @@ import io
 from flask import Blueprint, jsonify, request, current_app as app, g
 from werkzeug.datastructures import FileStorage
 
+import layman.common.rest
 from layman.http import LaymanError
 from layman.util import check_username_decorator, url_for
 from . import util, MAP_TYPE
@@ -112,7 +113,7 @@ def post(username):
             'actor_name': actor_name
         }
 
-        authz_util.setup_post_access_rights(request.form, kwargs, actor_name)
+        layman.common.rest.setup_post_access_rights(request.form, kwargs, actor_name)
 
         util.post_map(
             username,
